+++ src/ring_buffer/iter.rs
@@ -2,8 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use std::iter::FusedIterator;
-use std::marker::PhantomData;
+use core::iter::FusedIterator;
+use core::marker::PhantomData;
 
 use crate::types::ChunkLength;